+++ vizro-core/examples/scratch_dev/app.py
@@ -18,6 +18,7 @@ df_kpi = pd.DataFrame(
 home = vm.Page(
     title="Page Title",
     components=[
+        # kpi_card(data_frame=df_kpi, value_column="Actual", title="KPI with value"),
         vm.Figure(figure=kpi_card(data_frame=df_kpi, value_column="Actual", title="KPI with value")),
     ],
 )
